third_party/blink/web_tests/fast/replaced/table-percent-width-expected.txt

This test checks that a replaced element with percentage width (and no height specified) within a table cell is squeezed to the dimensions of the table cell.
See bug #29447.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS getWidth('img-1') is '105px'
PASS getHeight('img-1') is '105px'
PASS getWidth('img-2') is '100px'
PASS getHeight('img-2') is '100px'
PASS getWidth('img-3') is '40px'
PASS getHeight('img-3') is '40px'
PASS getWidth('img-4') is '100px'
PASS getHeight('img-4') is '100px'
PASS getWidth('img-5') is '40px'
PASS getHeight('img-5') is '40px'
PASS successfullyParsed is true

TEST COMPLETE
